requirements.txt
python-dotenv
dotenv
jwt
PyJWT
pyserial
RPi.GPIO

